separate restaurantEntry localStorage access from main.js
