+++ source/restaurantEntry/restaurantEntry.html
@@ -11,6 +11,7 @@
   
   <!-- Main Stylesheets & Scripts -->
   <link rel="stylesheet" href="main.css" />
+  <script src="./restaurantEntryRepo.js" type="module"></script>
   <script src="./main.js" type="module"></script>
 </head>
 <body>
